Minor changes to solve issues founded on Scrutinizer - [X] Reviewed "Best Practice" 18 issues - [X] Reviewed "Bug" 145 issues - [X] Reviewed "Comprehensibility" 3 issues - [X] Reviewed "Documentation" 1 issues - [X] Reviewed "Security" 3 issues - [ ] Reviewed "Unused Code" 14 issues
Trying to reduce issues on 74 files with issues.

+++ Core/App/AppInstaller.php
@@ -76,6 +76,7 @@ class AppInstaller
         } elseif ($installed) {
             header('Location: ' . $this->getUri());
         } elseif ('TRUE' === $this->request->get('phpinfo', '')) {
+            /** @noinspection ForgottenDebugOutputInspection */
             phpinfo();
         } else {
             $this->render();
@@ -136,7 +137,7 @@ class AppInstaller
             }
         }
 
-        chmod('Plugins', octdec(777));
+        chmod('Plugins', (int) octdec(777));
         $pluginManager = new PluginManager();
         $pluginManager->deploy();
         return true;
